pyproject.toml
[tool.poetry]
name = "api_yamdb"
version = "0.1.0"
description = ""
authors = ["SaPer663 <saper663ru@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "2.26.0"
Django = "2.2.16"
djangorestframework = "3.12.4"
PyJWT = "2.1.0"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "4.7.2"


[tool.poetry.dev-dependencies]
pytest = "6.2.4"
pytest-django = "4.4.0"
pytest-pythonpath = "0.7.3"
flake8 = "^4.0.1"
isort = "^5.10.1"
ipython = "^7.29.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
